+++ src/src/di_tile_map.cpp
@@ -33,7 +33,7 @@
 
 DiTileMap::DiTileMap(uint32_t screen_width, uint32_t screen_height,
                       uint32_t columns, uint32_t rows,
-                      uint32_t tile_width, uint32_t tile_height, uint8_t flags) {
+                      uint32_t tile_width, uint32_t tile_height, uint16_t flags) {
   m_tile_width = tile_width;
   m_tile_height = tile_height;
   m_rows = rows;
